python_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
websockets==12.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.3

